refactor FE, readme update, finalize testing

+++ spring-petclinic-visits-service/src/main/java/org/springframework/samples/petclinic/visit/domain/VisitRepository.java
@@ -7,7 +7,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface VisitRepository extends JpaRepository<Visit, Integer> {
+public interface VisitRepository extends JpaRepository<Visit, Long> {
 
     List<Visit> findByPetId(int petId);
